@@ -15,39 +15,41 @@
   },
   "devDependencies": {
     "babel-cli": "^6.24.1",
-    "babel-eslint": "^7.2.1",
-    "babel-jest": "^19.0.0",
+    "babel-eslint": "^7.2.3",
+    "babel-generator": "^6.24.1",
+    "babel-jest": "^20.0.3",
     "babel-plugin-external-helpers": "^6.22.0",
     "babel-plugin-syntax-jsx": "^6.18.0",
-    "babel-plugin-transform-class-properties": "^6.23.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-plugin-transform-remove-strict-mode": "^0.0.2",
-    "babel-preset-env": "^1.3.2",
-    "babel-preset-react": "^6.23.0",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-react": "^6.24.1",
     "chalk": "^1.1.3",
     "condition-circle": "^1.5.0",
     "cz-lerna-changelog": "^1.2.1",
-    "enzyme": "^2.8.0",
-    "enzyme-to-json": "^1.5.0",
+    "enzyme": "^2.8.2",
+    "enzyme-to-json": "^1.5.1",
     "eslint": "^3.19.0",
-    "eslint-config-standard": "^10.0.0",
-    "eslint-plugin-flowtype": "^2.30.4",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-plugin-flowtype": "^2.33.0",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-node": "^4.2.2",
     "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-react": "^6.10.3",
-    "eslint-plugin-standard": "^2.2.0",
-    "flow-bin": "^0.43.0",
-    "flow-typed": "^2.0.0",
+    "eslint-plugin-react": "^7.0.1",
+    "eslint-plugin-standard": "^3.0.1",
+    "flow-bin": "^0.46.0",
+    "flow-typed": "^2.1.2",
     "glob": "^7.1.1",
     "jasmine-reporters": "^2.2.1",
-    "jest": "^19.0.2",
-    "lerna": "^2.0.0-beta.38",
-    "lerna-semantic-release": "^9.0.7",
+    "jest": "^20.0.3",
+    "lerna": "^2.0.0-rc.4",
+    "lerna-semantic-release": "^9.1.0",
     "node-emoji": "^1.5.1",
-    "react": "^15.4.2",
-    "react-addons-test-utils": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "react": "^15.5.4",
+    "react-addons-test-utils": "^15.5.1",
+    "react-dom": "^15.5.4",
+    "react-test-renderer": "^15.5.4",
     "rimraf": "^2.6.1",
     "tmp": "^0.0.31"
   },
