@@ -12,19 +12,19 @@
     "start": "webpack-dev-server --hot"
   },
   "dependencies": {
-    "lingui-cli": "^0.11.0",
-    "lingui-react": "^0.6.0",
+    "lingui-cli": "^0.12.0",
+    "lingui-react": "^0.8.0",
     "react": "15.x"
   },
   "devDependencies": {
     "babel-cli": "^6.24.1",
-    "babel-loader": "^6.4.1",
+    "babel-loader": "^7.0.0",
     "babel-polyfill": "^6.23.0",
-    "babel-preset-env": "^1.3.3",
-    "babel-preset-lingui-react": "^0.3.0",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-lingui-react": "^0.4.0",
     "babel-preset-react": "^6.24.1",
-    "webpack": "^2.3.3",
-    "webpack-dev-server": "^2.4.2"
+    "webpack": "^2.5.1",
+    "webpack-dev-server": "^2.4.5"
   },
   "babel": {
     "presets": [
