@@ -1,5 +1,5 @@
-// flow-typed signature: 634257d6ff0cd9119a73a3c52f3940d2
-// flow-typed version: <<STUB>>/babel-cli_v^6.24.1/flow_v0.43.0
+// flow-typed signature: 8615cf6c39596ff7ba27c4a0377fe299
+// flow-typed version: <<STUB>>/babel-cli_v^6.24.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
