@@ -1,5 +1,5 @@
-// flow-typed signature: 743e80cdbc020f20b680e9125460683f
-// flow-typed version: <<STUB>>/babel-eslint_v^7.2.1/flow_v0.43.0
+// flow-typed signature: f3c3dd3c4bea413962ad456458ec0804
+// flow-typed version: <<STUB>>/babel-eslint_v^7.2.3/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
@@ -23,6 +23,10 @@ declare module 'babel-eslint' {
  * needed.
  */
 declare module 'babel-eslint/babylon-to-espree/attachComments' {
+  declare module.exports: any;
+}
+
+declare module 'babel-eslint/babylon-to-espree/convertComments' {
   declare module.exports: any;
 }
 
@@ -49,6 +53,9 @@ declare module 'babel-eslint/babylon-to-espree/toTokens' {
 // Filename aliases
 declare module 'babel-eslint/babylon-to-espree/attachComments.js' {
   declare module.exports: $Exports<'babel-eslint/babylon-to-espree/attachComments'>;
+}
+declare module 'babel-eslint/babylon-to-espree/convertComments.js' {
+  declare module.exports: $Exports<'babel-eslint/babylon-to-espree/convertComments'>;
 }
 declare module 'babel-eslint/babylon-to-espree/convertTemplateType.js' {
   declare module.exports: $Exports<'babel-eslint/babylon-to-espree/convertTemplateType'>;
