@@ -1,5 +1,5 @@
-// flow-typed signature: 71591a897e9fca2a6b850c5d1a4b1f86
-// flow-typed version: <<STUB>>/babel-jest_v^19.0.0/flow_v0.43.0
+// flow-typed signature: d4e78a10db1c03409f039c1842f09fc1
+// flow-typed version: <<STUB>>/babel-jest_v^20.0.3/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
