@@ -1,5 +1,5 @@
-// flow-typed signature: babdd38cac1634b4e13fd1391144e1c7
-// flow-typed version: <<STUB>>/babel-plugin-external-helpers_v^6.22.0/flow_v0.43.0
+// flow-typed signature: deffac357ce63a592e2d820eeb12a55c
+// flow-typed version: <<STUB>>/babel-plugin-external-helpers_v^6.22.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
