@@ -1,5 +1,5 @@
-// flow-typed signature: efa777c134212551df885830d561dfe4
-// flow-typed version: <<STUB>>/babel-plugin-syntax-jsx_v^6.18.0/flow_v0.43.0
+// flow-typed signature: c4168fd4b4ed39f8b367b2a5ead74580
+// flow-typed version: <<STUB>>/babel-plugin-syntax-jsx_v^6.18.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
