@@ -1,5 +1,5 @@
-// flow-typed signature: 77b28d719960c8a6f245dbe0624f1bb8
-// flow-typed version: <<STUB>>/babel-plugin-transform-class-properties_v^6.23.0/flow_v0.43.0
+// flow-typed signature: 805d3c8609e612f9f5eeb61e6e4d52ee
+// flow-typed version: <<STUB>>/babel-plugin-transform-class-properties_v^6.24.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
