@@ -1,5 +1,5 @@
-// flow-typed signature: c09f74e06735d64336957d71937485fd
-// flow-typed version: <<STUB>>/babel-plugin-transform-remove-strict-mode_v^0.0.2/flow_v0.43.0
+// flow-typed signature: 2e91b96b4e6d42ff07d32da88e02c382
+// flow-typed version: <<STUB>>/babel-plugin-transform-remove-strict-mode_v^0.0.2/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
