@@ -1,5 +1,5 @@
-// flow-typed signature: 04d8a379137a6ca5f200a654bfc0b064
-// flow-typed version: <<STUB>>/babel-preset-env_v^1.3.2/flow_v0.43.0
+// flow-typed signature: f12875965303032e2f229e46f5734514
+// flow-typed version: <<STUB>>/babel-preset-env_v^1.4.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
