@@ -1,5 +1,5 @@
-// flow-typed signature: 8c0c58f53dba0f919af55aa788f28f3b
-// flow-typed version: <<STUB>>/babel-preset-react_v^6.23.0/flow_v0.43.0
+// flow-typed signature: f36cef8c0faabcd12873f8dedbba14fd
+// flow-typed version: <<STUB>>/babel-preset-react_v^6.24.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
