@@ -1,5 +1,5 @@
-// flow-typed signature: 22163d5ebedee56fefd6ade37d970548
-// flow-typed version: <<STUB>>/condition-circle_v^1.5.0/flow_v0.43.0
+// flow-typed signature: eb32bf635bb16316c512f4573f71ec0e
+// flow-typed version: <<STUB>>/condition-circle_v^1.5.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
