@@ -1,5 +1,5 @@
-// flow-typed signature: c4ab6bca699ff967b143b1a7bd1d9fba
-// flow-typed version: <<STUB>>/cz-lerna-changelog_v^1.2.1/flow_v0.43.0
+// flow-typed signature: b8099df95af3fc19911964d9200ccebb
+// flow-typed version: <<STUB>>/cz-lerna-changelog_v^1.2.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
