@@ -1,5 +1,5 @@
-// flow-typed signature: a157e6cd3816f703c307e150d63886ba
-// flow-typed version: <<STUB>>/enzyme-to-json_v^1.5.0/flow_v0.43.0
+// flow-typed signature: d46ada5d05959b8ec66d2f86890b7338
+// flow-typed version: <<STUB>>/enzyme-to-json_v^1.5.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
