@@ -1,5 +1,5 @@
-// flow-typed signature: 467b526579c54e371491a416a688644a
-// flow-typed version: <<STUB>>/eslint-config-standard_v^10.0.0/flow_v0.43.0
+// flow-typed signature: b279cfa33de1526b530d77d718443a1b
+// flow-typed version: <<STUB>>/eslint-config-standard_v^10.2.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
