@@ -1,5 +1,5 @@
-// flow-typed signature: dd6d091df8fa472025ea91de7e7e8171
-// flow-typed version: <<STUB>>/eslint-plugin-flowtype_v^2.30.4/flow_v0.43.0
+// flow-typed signature: d3a72d7b59d5bd2824d44ae2723bbb86
+// flow-typed version: <<STUB>>/eslint-plugin-flowtype_v^2.33.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
@@ -51,6 +51,10 @@ declare module 'eslint-plugin-flowtype/dist/rules/noDupeKeys' {
 }
 
 declare module 'eslint-plugin-flowtype/dist/rules/noPrimitiveConstructorTypes' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/noTypesMissingFileAnnotation' {
   declare module.exports: any;
 }
 
@@ -214,6 +218,9 @@ declare module 'eslint-plugin-flowtype/dist/rules/noDupeKeys.js' {
 }
 declare module 'eslint-plugin-flowtype/dist/rules/noPrimitiveConstructorTypes.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/noPrimitiveConstructorTypes'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/noTypesMissingFileAnnotation.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/noTypesMissingFileAnnotation'>;
 }
 declare module 'eslint-plugin-flowtype/dist/rules/noWeakTypes.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/noWeakTypes'>;
