@@ -1,5 +1,5 @@
-// flow-typed signature: 07f7a987d1c87deb4f730a58815c149d
-// flow-typed version: <<STUB>>/eslint-plugin-import_v^2.2.0/flow_v0.43.0
+// flow-typed signature: fdc069572f193f76e11f3ff6be5fde49
+// flow-typed version: <<STUB>>/eslint-plugin-import_v^2.2.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
