@@ -1,5 +1,5 @@
-// flow-typed signature: d6e48fc314bd208d45c41a76912ba069
-// flow-typed version: <<STUB>>/eslint-plugin-node_v^4.2.2/flow_v0.43.0
+// flow-typed signature: 443da165bb759d374c0c461c3b0dc6d6
+// flow-typed version: <<STUB>>/eslint-plugin-node_v^4.2.2/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
