@@ -1,5 +1,5 @@
-// flow-typed signature: b30be331f59f6688f6f2e0ac24bd225b
-// flow-typed version: <<STUB>>/eslint-plugin-promise_v^3.5.0/flow_v0.43.0
+// flow-typed signature: ff44d9ca4a2b69c17a28bf86aa17c310
+// flow-typed version: <<STUB>>/eslint-plugin-promise_v^3.5.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
