@@ -1,5 +1,5 @@
-// flow-typed signature: d8c6c3f8b64b4a9d2542098354a4ca97
-// flow-typed version: <<STUB>>/eslint-plugin-react_v^6.10.3/flow_v0.43.0
+// flow-typed signature: bcd88766b0f6ace18b0f3b08fc38c2df
+// flow-typed version: <<STUB>>/eslint-plugin-react_v^7.0.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
@@ -142,10 +142,6 @@ declare module 'eslint-plugin-react/lib/rules/no-array-index-key' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-react/lib/rules/no-comment-textnodes' {
-  declare module.exports: any;
-}
-
 declare module 'eslint-plugin-react/lib/rules/no-danger-with-children' {
   declare module.exports: any;
 }
@@ -210,6 +206,10 @@ declare module 'eslint-plugin-react/lib/rules/no-unused-prop-types' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/rules/no-will-update-set-state' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/rules/prefer-es6-class' {
   declare module.exports: any;
 }
@@ -227,10 +227,6 @@ declare module 'eslint-plugin-react/lib/rules/react-in-jsx-scope' {
 }
 
 declare module 'eslint-plugin-react/lib/rules/require-default-props' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-react/lib/rules/require-extension' {
   declare module.exports: any;
 }
 
@@ -262,10 +258,6 @@ declare module 'eslint-plugin-react/lib/rules/void-dom-elements-no-children' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-react/lib/rules/wrap-multilines' {
-  declare module.exports: any;
-}
-
 declare module 'eslint-plugin-react/lib/util/annotations' {
   declare module.exports: any;
 }
@@ -275,6 +267,10 @@ declare module 'eslint-plugin-react/lib/util/Components' {
 }
 
 declare module 'eslint-plugin-react/lib/util/getTokenBeforeClosingBracket' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/makeNoMethodSetStateRule' {
   declare module.exports: any;
 }
 
@@ -387,9 +383,6 @@ declare module 'eslint-plugin-react/lib/rules/jsx-wrap-multilines.js' {
 declare module 'eslint-plugin-react/lib/rules/no-array-index-key.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-array-index-key'>;
 }
-declare module 'eslint-plugin-react/lib/rules/no-comment-textnodes.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-comment-textnodes'>;
-}
 declare module 'eslint-plugin-react/lib/rules/no-danger-with-children.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-danger-with-children'>;
 }
@@ -438,6 +431,9 @@ declare module 'eslint-plugin-react/lib/rules/no-unknown-property.js' {
 declare module 'eslint-plugin-react/lib/rules/no-unused-prop-types.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-unused-prop-types'>;
 }
+declare module 'eslint-plugin-react/lib/rules/no-will-update-set-state.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-will-update-set-state'>;
+}
 declare module 'eslint-plugin-react/lib/rules/prefer-es6-class.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/prefer-es6-class'>;
 }
@@ -452,9 +448,6 @@ declare module 'eslint-plugin-react/lib/rules/react-in-jsx-scope.js' {
 }
 declare module 'eslint-plugin-react/lib/rules/require-default-props.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/require-default-props'>;
-}
-declare module 'eslint-plugin-react/lib/rules/require-extension.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/require-extension'>;
 }
 declare module 'eslint-plugin-react/lib/rules/require-optimization.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/require-optimization'>;
@@ -477,9 +470,6 @@ declare module 'eslint-plugin-react/lib/rules/style-prop-object.js' {
 declare module 'eslint-plugin-react/lib/rules/void-dom-elements-no-children.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/void-dom-elements-no-children'>;
 }
-declare module 'eslint-plugin-react/lib/rules/wrap-multilines.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/wrap-multilines'>;
-}
 declare module 'eslint-plugin-react/lib/util/annotations.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/annotations'>;
 }
@@ -488,6 +478,9 @@ declare module 'eslint-plugin-react/lib/util/Components.js' {
 }
 declare module 'eslint-plugin-react/lib/util/getTokenBeforeClosingBracket.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/getTokenBeforeClosingBracket'>;
+}
+declare module 'eslint-plugin-react/lib/util/makeNoMethodSetStateRule.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/makeNoMethodSetStateRule'>;
 }
 declare module 'eslint-plugin-react/lib/util/pragma.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/pragma'>;
