@@ -1,5 +1,5 @@
-// flow-typed signature: 4997eedfc77001a191af3a2153c958f6
-// flow-typed version: <<STUB>>/eslint-plugin-standard_v^2.2.0/flow_v0.43.0
+// flow-typed signature: 75f7fc712c6e3acd0ff1b6166e8ed9b9
+// flow-typed version: <<STUB>>/eslint-plugin-standard_v^3.0.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
