@@ -1,5 +1,5 @@
-// flow-typed signature: cb6fe5abf713c96141603e00eb35fcd4
-// flow-typed version: <<STUB>>/eslint_v^3.19.0/flow_v0.43.0
+// flow-typed signature: 680e2c3df27f3b57842d2164c5fec929
+// flow-typed version: <<STUB>>/eslint_v^3.19.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
