@@ -1,5 +1,5 @@
-// flow-typed signature: 806f44f9be8f8a70d639d88eeab13ba7
-// flow-typed version: <<STUB>>/flow-typed_v^2.0.0/flow_v0.43.0
+// flow-typed signature: 8898603ed592b58f9cb89fd30a371a09
+// flow-typed version: <<STUB>>/flow-typed_v^2.1.2/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
@@ -58,6 +58,10 @@ declare module 'flow-typed/dist/commands/version' {
   declare module.exports: any;
 }
 
+declare module 'flow-typed/dist/lib/cacheRepoUtils' {
+  declare module.exports: any;
+}
+
 declare module 'flow-typed/dist/lib/codeSign' {
   declare module.exports: any;
 }
@@ -70,11 +74,19 @@ declare module 'flow-typed/dist/lib/flowProjectUtils' {
   declare module.exports: any;
 }
 
+declare module 'flow-typed/dist/lib/flowVersion' {
+  declare module.exports: any;
+}
+
 declare module 'flow-typed/dist/lib/git' {
   declare module.exports: any;
 }
 
 declare module 'flow-typed/dist/lib/github' {
+  declare module.exports: any;
+}
+
+declare module 'flow-typed/dist/lib/isInFlowTypedRepo' {
   declare module.exports: any;
 }
 
@@ -86,7 +98,11 @@ declare module 'flow-typed/dist/lib/node' {
   declare module.exports: any;
 }
 
-declare module 'flow-typed/dist/lib/npmProjectUtils' {
+declare module 'flow-typed/dist/lib/npm/npmLibDefs' {
+  declare module.exports: any;
+}
+
+declare module 'flow-typed/dist/lib/npm/npmProjectUtils' {
   declare module.exports: any;
 }
 
@@ -95,6 +111,10 @@ declare module 'flow-typed/dist/lib/semver' {
 }
 
 declare module 'flow-typed/dist/lib/stubUtils' {
+  declare module.exports: any;
+}
+
+declare module 'flow-typed/dist/lib/validationErrors' {
   declare module.exports: any;
 }
 
@@ -126,6 +146,9 @@ declare module 'flow-typed/dist/commands/validateDefs.js' {
 declare module 'flow-typed/dist/commands/version.js' {
   declare module.exports: $Exports<'flow-typed/dist/commands/version'>;
 }
+declare module 'flow-typed/dist/lib/cacheRepoUtils.js' {
+  declare module.exports: $Exports<'flow-typed/dist/lib/cacheRepoUtils'>;
+}
 declare module 'flow-typed/dist/lib/codeSign.js' {
   declare module.exports: $Exports<'flow-typed/dist/lib/codeSign'>;
 }
@@ -135,11 +158,17 @@ declare module 'flow-typed/dist/lib/fileUtils.js' {
 declare module 'flow-typed/dist/lib/flowProjectUtils.js' {
   declare module.exports: $Exports<'flow-typed/dist/lib/flowProjectUtils'>;
 }
+declare module 'flow-typed/dist/lib/flowVersion.js' {
+  declare module.exports: $Exports<'flow-typed/dist/lib/flowVersion'>;
+}
 declare module 'flow-typed/dist/lib/git.js' {
   declare module.exports: $Exports<'flow-typed/dist/lib/git'>;
 }
 declare module 'flow-typed/dist/lib/github.js' {
   declare module.exports: $Exports<'flow-typed/dist/lib/github'>;
+}
+declare module 'flow-typed/dist/lib/isInFlowTypedRepo.js' {
+  declare module.exports: $Exports<'flow-typed/dist/lib/isInFlowTypedRepo'>;
 }
 declare module 'flow-typed/dist/lib/libDefs.js' {
   declare module.exports: $Exports<'flow-typed/dist/lib/libDefs'>;
@@ -147,12 +176,18 @@ declare module 'flow-typed/dist/lib/libDefs.js' {
 declare module 'flow-typed/dist/lib/node.js' {
   declare module.exports: $Exports<'flow-typed/dist/lib/node'>;
 }
-declare module 'flow-typed/dist/lib/npmProjectUtils.js' {
-  declare module.exports: $Exports<'flow-typed/dist/lib/npmProjectUtils'>;
+declare module 'flow-typed/dist/lib/npm/npmLibDefs.js' {
+  declare module.exports: $Exports<'flow-typed/dist/lib/npm/npmLibDefs'>;
+}
+declare module 'flow-typed/dist/lib/npm/npmProjectUtils.js' {
+  declare module.exports: $Exports<'flow-typed/dist/lib/npm/npmProjectUtils'>;
 }
 declare module 'flow-typed/dist/lib/semver.js' {
   declare module.exports: $Exports<'flow-typed/dist/lib/semver'>;
 }
 declare module 'flow-typed/dist/lib/stubUtils.js' {
   declare module.exports: $Exports<'flow-typed/dist/lib/stubUtils'>;
+}
+declare module 'flow-typed/dist/lib/validationErrors.js' {
+  declare module.exports: $Exports<'flow-typed/dist/lib/validationErrors'>;
 }
