@@ -1,5 +1,5 @@
-// flow-typed signature: 91441b745ffe6b6563790220cf0b7f21
-// flow-typed version: <<STUB>>/glob_v^7.1.1/flow_v0.43.0
+// flow-typed signature: c54b5dbcd2c8c56943e6aabe2a927230
+// flow-typed version: <<STUB>>/glob_v^7.1.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
