@@ -1,5 +1,5 @@
-// flow-typed signature: 8955eec195c140286360dc8d572e4ad1
-// flow-typed version: <<STUB>>/jasmine-reporters_v^2.2.1/flow_v0.43.0
+// flow-typed signature: ed335c01295c3df90578bd57b7a16bc2
+// flow-typed version: <<STUB>>/jasmine-reporters_v^2.2.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
