@@ -1,5 +1,5 @@
-// flow-typed signature: b3ed97c44539e6cdbaf9032b315a2b31
-// flow-typed version: ea7ac31527/jest_v19.x.x/flow_>=v0.33.x
+// flow-typed signature: eb392bde0440325fce2421d18447acde
+// flow-typed version: 0786be2fed/jest_v20.x.x/flow_>=v0.33.x
 
 type JestMockFn = {
   (...args: Array<any>): any,
@@ -17,7 +17,7 @@ type JestMockFn = {
      * An array that contains all the object instances that have been
      * instantiated from this mock function.
      */
-    instances: mixed,
+    instances: mixed
   },
   /**
    * Resets all information stored in the mockFn.mock.calls and
@@ -54,15 +54,15 @@ type JestMockFn = {
   /**
    * Sugar for only returning a value once inside your mock
    */
-  mockReturnValueOnce(value: any): JestMockFn,
-}
+  mockReturnValueOnce(value: any): JestMockFn
+};
 
 type JestAsymmetricEqualityType = {
   /**
    * A custom Jasmine equality tester
    */
-  asymmetricMatch(value: mixed): boolean,
-}
+  asymmetricMatch(value: mixed): boolean
+};
 
 type JestCallsType = {
   allArgs(): mixed,
@@ -71,22 +71,35 @@ type JestCallsType = {
   count(): number,
   first(): mixed,
   mostRecent(): mixed,
-  reset(): void,
-}
+  reset(): void
+};
 
 type JestClockType = {
   install(): void,
   mockDate(date: Date): void,
   tick(): void,
-  uninstall(): void,
-}
+  uninstall(): void
+};
 
 type JestMatcherResult = {
-  message?: string | ()=>string,
-  pass: boolean,
-}
+  message?: string | (() => string),
+  pass: boolean
+};
 
 type JestMatcher = (actual: any, expected: any) => JestMatcherResult;
+
+type JestPromiseType = {
+  /**
+   * Use rejects to unwrap the reason of a rejected promise so any other
+   * matcher can be chained. If the promise is fulfilled the assertion fails.
+   */
+  rejects: JestExpectType,
+  /**
+   * Use resolves to unwrap the value of a fulfilled promise so any other
+   * matcher can be chained. If the promise is rejected the assertion fails.
+   */
+  resolves: JestExpectType
+};
 
 type JestExpectType = {
   not: JestExpectType,
@@ -224,8 +237,8 @@ type JestExpectType = {
    * Use .toThrowErrorMatchingSnapshot to test that a function throws a error
    * matching the most recent snapshot when it is called.
    */
-  toThrowErrorMatchingSnapshot(): void,
-}
+  toThrowErrorMatchingSnapshot(): void
+};
 
 type JestObjectType = {
   /**
@@ -351,12 +364,12 @@ type JestObjectType = {
    * Creates a mock function similar to jest.fn but also tracks calls to
    * object[methodName].
    */
-  spyOn(object: Object, methodName: string): JestMockFn,
-}
+  spyOn(object: Object, methodName: string): JestMockFn
+};
 
 type JestSpyType = {
-  calls: JestCallsType,
-}
+  calls: JestCallsType
+};
 
 /** Runs this function after every test inside this context */
 declare function afterEach(fn: Function): void;
@@ -398,7 +411,7 @@ declare var it: {
    * @param {string} Name of Test
    * @param {Function} Test
    */
-  concurrent(name: string, fn?: Function): ?Promise<void>,
+  concurrent(name: string, fn?: Function): ?Promise<void>
 };
 declare function fit(name: string, fn: Function): ?Promise<void>;
 /** An individual test unit */
@@ -415,19 +428,20 @@ declare var xtest: typeof it;
 /** The expect function is used every time you want to test a value */
 declare var expect: {
   /** The object that you want to make assertions against */
-  (value: any): JestExpectType,
+  (value: any): JestExpectType & JestPromiseType,
   /** Add additional Jasmine matchers to Jest's roster */
-  extend(matchers: {[name:string]: JestMatcher}): void,
+  extend(matchers: { [name: string]: JestMatcher }): void,
   /** Add a module that formats application-specific data structures. */
   addSnapshotSerializer(serializer: (input: Object) => string): void,
   assertions(expectedAssertions: number): void,
+  hasAssertions(): void,
   any(value: mixed): JestAsymmetricEqualityType,
   anything(): void,
   arrayContaining(value: Array<mixed>): void,
   objectContaining(value: Object): void,
   /** Matches any received string that contains the exact expected string. */
   stringContaining(value: string): void,
-  stringMatching(value: string | RegExp): void,
+  stringMatching(value: string | RegExp): void
 };
 
 // TODO handle return type
@@ -435,7 +449,7 @@ declare var expect: {
 declare function spyOn(value: mixed, method: string): Object;
 
 /** Holds all functions related to manipulating test runner */
-declare var jest: JestObjectType
+declare var jest: JestObjectType;
 
 /**
  * The global Jamine object, this is generally not exposed as the public API,
@@ -448,7 +462,10 @@ declare var jasmine: {
   arrayContaining(value: Array<mixed>): void,
   clock(): JestClockType,
   createSpy(name: string): JestSpyType,
-  createSpyObj(baseName: string, methodNames: Array<string>): {[methodName: string]: JestSpyType},
+  createSpyObj(
+    baseName: string,
+    methodNames: Array<string>
+  ): { [methodName: string]: JestSpyType },
   objectContaining(value: Object): void,
-  stringMatching(value: string): void,
-}
+  stringMatching(value: string): void
+};
