@@ -1,5 +1,5 @@
-// flow-typed signature: cf4c3300745ba4ae0ea0cd436d8f8afe
-// flow-typed version: <<STUB>>/lerna-semantic-release_v^9.0.7/flow_v0.43.0
+// flow-typed signature: c8aa1ca9531c1a028c3f22753a5c095b
+// flow-typed version: <<STUB>>/lerna-semantic-release_v^9.1.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
