@@ -1,5 +1,5 @@
-// flow-typed signature: 6f7d5e7104c5ef06f9d9f0e2160e2c90
-// flow-typed version: <<STUB>>/lerna_v^2.0.0-beta.38/flow_v0.43.0
+// flow-typed signature: e6ef0f273d2dbe2d392a9cfcc843fa71
+// flow-typed version: <<STUB>>/lerna_v^2.0.0-rc.4/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
@@ -23,6 +23,10 @@ declare module 'lerna' {
  * needed.
  */
 declare module 'lerna/bin/lerna' {
+  declare module.exports: any;
+}
+
+declare module 'lerna/lib/cli' {
   declare module.exports: any;
 }
 
@@ -67,6 +71,10 @@ declare module 'lerna/lib/commands/RunCommand' {
 }
 
 declare module 'lerna/lib/commands/UpdatedCommand' {
+  declare module.exports: any;
+}
+
+declare module 'lerna/lib/ConventionalCommitUtilities' {
   declare module.exports: any;
 }
 
@@ -126,9 +134,28 @@ declare module 'lerna/lib/UpdatedPackagesCollector' {
   declare module.exports: any;
 }
 
+declare module 'lerna/lib/utils/dependencyIsSatisfied' {
+  declare module.exports: any;
+}
+
+declare module 'lerna/lib/utils/filterFlags' {
+  declare module.exports: any;
+}
+
+declare module 'lerna/lib/utils/output' {
+  declare module.exports: any;
+}
+
+declare module 'lerna/lib/utils/writeLogFile' {
+  declare module.exports: any;
+}
+
 // Filename aliases
 declare module 'lerna/bin/lerna.js' {
   declare module.exports: $Exports<'lerna/bin/lerna'>;
+}
+declare module 'lerna/lib/cli.js' {
+  declare module.exports: $Exports<'lerna/lib/cli'>;
 }
 declare module 'lerna/lib/Command.js' {
   declare module.exports: $Exports<'lerna/lib/Command'>;
@@ -162,6 +189,9 @@ declare module 'lerna/lib/commands/RunCommand.js' {
 }
 declare module 'lerna/lib/commands/UpdatedCommand.js' {
   declare module.exports: $Exports<'lerna/lib/commands/UpdatedCommand'>;
+}
+declare module 'lerna/lib/ConventionalCommitUtilities.js' {
+  declare module.exports: $Exports<'lerna/lib/ConventionalCommitUtilities'>;
 }
 declare module 'lerna/lib/ExitHandler.js' {
   declare module.exports: $Exports<'lerna/lib/ExitHandler'>;
@@ -204,4 +234,16 @@ declare module 'lerna/lib/Repository.js' {
 }
 declare module 'lerna/lib/UpdatedPackagesCollector.js' {
   declare module.exports: $Exports<'lerna/lib/UpdatedPackagesCollector'>;
+}
+declare module 'lerna/lib/utils/dependencyIsSatisfied.js' {
+  declare module.exports: $Exports<'lerna/lib/utils/dependencyIsSatisfied'>;
+}
+declare module 'lerna/lib/utils/filterFlags.js' {
+  declare module.exports: $Exports<'lerna/lib/utils/filterFlags'>;
+}
+declare module 'lerna/lib/utils/output.js' {
+  declare module.exports: $Exports<'lerna/lib/utils/output'>;
+}
+declare module 'lerna/lib/utils/writeLogFile.js' {
+  declare module.exports: $Exports<'lerna/lib/utils/writeLogFile'>;
 }
