@@ -1,5 +1,5 @@
-// flow-typed signature: f23ddcbd25cc7faeab51da4ae97d7236
-// flow-typed version: <<STUB>>/node-emoji_v^1.5.1/flow_v0.43.0
+// flow-typed signature: 770e63ebdf3602dc38e73d1526706ab0
+// flow-typed version: <<STUB>>/node-emoji_v^1.5.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
