@@ -1,5 +1,5 @@
-// flow-typed signature: 037358386065a9822dc62ed593b874a1
-// flow-typed version: <<STUB>>/rimraf_v^2.6.1/flow_v0.43.0
+// flow-typed signature: 475d59277bfba3e04a313defb589ae39
+// flow-typed version: <<STUB>>/rimraf_v^2.6.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
