@@ -1,5 +1,5 @@
-// flow-typed signature: c464ccb2566b238baec75370234bddd1
-// flow-typed version: <<STUB>>/tmp_v^0.0.31/flow_v0.43.0
+// flow-typed signature: fd056be283ec860b60961e61493797e7
+// flow-typed version: <<STUB>>/tmp_v^0.0.31/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -8,8 +8,8 @@
  *
  * Fill this stub out by replacing all the `any` types.
  *
- * Once filled out, we encourage you to share your work with the 
- * community by sending a pull request to: 
+ * Once filled out, we encourage you to share your work with the
+ * community by sending a pull request to:
  * https://github.com/flowtype/flow-typed
  */
 
